+++ zipline/transforms/vwap.py
@@ -17,10 +17,10 @@
 from numbers import Number
 from collections import defaultdict
 
-from zipline.gens.transform import EventWindow, TransformMeta
+from zipline.transforms.utils import EventWindow, TransformMeta
 
 
-class VWAP(object):
+class MovingVWAP(object):
     """
     Class that maintains a dictionary from sids to VWAPEventWindows.
     """
